chore: refactor icon component files
