+++ meteor/client/lib/notifications/NotificationCenterPanel.tsx
@@ -9,7 +9,7 @@ import { ContextMenuTrigger, ContextMenu, MenuItem } from '@jstarpl/react-contex
 import * as _ from 'underscore'
 import { RundownId } from '../../../lib/collections/Rundowns'
 import { SegmentId } from '../../../lib/collections/Segments'
-import { CriticalIcon, WarningIcon, CollapseChevrons, InformationIcon } from '../notificationIcons'
+import { CriticalIcon, WarningIcon, CollapseChevrons, InformationIcon } from '../ui/icons/notifications'
 import update from 'immutability-helper'
 
 interface IPopUpProps {
