+++ meteor/client/lib/ui/icons/switchboard.tsx
