+++ meteor/client/ui/SegmentTimeline/PieceStatusIcon.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { NoticeLevel } from '../../lib/notifications/notifications'
-import { CriticalIconSmall, WarningIconSmall } from '../../lib/notificationIcons'
+import { CriticalIconSmall, WarningIconSmall } from '../../lib/ui/icons/notifications'
 
 export interface IProps {
 	noticeLevel: NoticeLevel
